+++ CHANGELOG.md
@@ -8,8 +8,8 @@
 - Added the `craft.blitz.dynamicInclude()` template variable, that includes a template via an AJAX request. The `dynamicInclude()` method always returns a freshly rendered template.
 - Added the `craft.blitz.fetch()` template variable, that fetches a URI via an AJAX request. Whether the URI response is cached or not is determined by the URI patterns in the plugin settings.
 - Added the `blitz/templates/include` and `blitz/templates/dynamic-include` controller actions.
-- Added the `ssiEnabled` config setting that enables Blitz to include templates using Server Side Includes (SSI), which must be enabled on the web server.
-- Added the `esiEnabled` config setting that enables Blitz to include templates using Edge Side Includes (ESI), which must be enabled on the web server or reverse proxy (CDN).
+- Added the `ssiEnabled` config setting that enables Blitz to include templates using Server-Side Includes (SSI), which must be enabled on the web server.
+- Added the `esiEnabled` config setting that enables Blitz to include templates using Edge-Side Includes (ESI), which must be enabled on the web server or reverse proxy (CDN).
 - Added the `maxUriLength` config setting.
 - Added a cached includes column to the Blitz cache utility for the File Cache Storage driver.